requirements.txt
pyjwt
pillow
flask
flask_sqlalchemy 
